Fix double to float conversion warning

+++ lre3/src/bindings/lre3_camera_binding.cpp
@@ -13,7 +13,7 @@ FBIND(LRE3Camera, SetZoom)
 {
     GET_SELF(LRE3Camera);
     GET_NUMBER(zoom, 2);
-    self->SetZoom(zoom);
+    self->SetZoom((float)zoom);
     return 0;
 }
